use proper name for created cra files in cloud, add a safety check when deleting a file

+++ droidbones/src/main/java/com/github/zibnix/droidbones/mvvm/FileService.kt
@@ -107,6 +107,7 @@ object FileService {
 
     fun deleteFile(file: File): Result<Unit> {
         return try {
+            if (!file.exists()) return Result.success(Unit)
             if (file.delete()) {
                 Result.success(Unit)
             } else {
